backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic
pydantic-settings
httpx
feedparser
apscheduler
python-dotenv
openai
pytest
pytest-asyncio
